pmbus: Update bus type in createPMBus() to uint16_t
Change the bus parameter of PMBus::createPMBus() from uint8_t to
uint16_t. Update all phosphor-power-supply calls accordingly. This
change required to support multi-chassis system, where the bus
identifier may exceed the range of uint8_t.

Tested:
    Simulated the bus ID greater than 255 and verified the apps works as
    expected.

+++ phosphor-power-supply/new_power_supply.hpp
@@ -91,7 +91,7 @@ class PowerSupply
      * @param[in] callback - Get the power on status of the psu manager class
      */
     PowerSupply(sdbusplus::bus_t& bus, const std::string& invpath,
-                std::uint8_t i2cbus, const std::uint16_t i2caddr,
+                std::uint16_t i2cbus, const std::uint16_t i2caddr,
                 const std::string& driver, const std::string& gpioLineName,
                 std::function<bool()>&& callback);
 
@@ -106,7 +106,7 @@ class PowerSupply
      * @param[in] chassisShortName - Chassis name
      */
     PowerSupply(sdbusplus::bus_t& bus, const std::string& invpath,
-                std::uint8_t i2cbus, const std::uint16_t i2caddr,
+                std::uint16_t i2cbus, const std::uint16_t i2caddr,
                 const std::string& driver, const std::string& gpioLineName,
                 std::function<bool()>&& callback,
                 const std::string& chassisShortName);
