+++ phosphor-power-supply/test/mock.cpp
@@ -5,7 +5,7 @@ namespace phosphor
 namespace pmbus
 {
 
-std::unique_ptr<PMBusBase> createPMBus(std::uint8_t /*bus*/,
+std::unique_ptr<PMBusBase> createPMBus(std::uint16_t /*bus*/,
                                        const std::string& /*address*/)
 {
     return std::make_unique<MockedPMBus>();
